changed the language of all .html added main controller

+++ racecup/src/main/java/com/epam/racecup/dao/RaceRepository.java
@@ -3,5 +3,5 @@ package com.epam.racecup.dao;
 import com.epam.racecup.model.Race;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RaceRepository extends JpaRepository<Race, Integer> {
+public interface RaceRepository extends JpaRepository<Race, Long> {
 }
